requirements.txt
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
coverage==5.3
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-pytest==0.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.1.0
Faker==4.17.1
flake8==3.8.4
importlib-metadata==3.1.0
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-flake8==1.0.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
zipp==3.4.0
